WIP: media: meson: vdec: commonize some HEVC code
Will be shared with the VP9 codebase.

%% original patch: 0212-WIP-media-meson-vdec-commonize-some-HEVC-code.patch
